Task 1 done

+++ lab2/pintos/src/threads/thread.h
@@ -89,8 +89,9 @@ struct thread
     uint8_t *stack;                     /* Saved stack pointer. */
     int priority;                       /* Priority. */
     struct list_elem allelem;           /* List element for all threads list. */
-    int time_wait;                      /* Number of ticks to wait */
-    int start_time;                     /* Starting time of sleep */
+//    int64_t end_time;
+    int64_t time_wait;                      /* Number of ticks to wait */
+    int64_t start_time;                     /* Starting time of sleep */
 
     /* Shared between thread.c and synch.c. */
     struct list_elem elem;              /* List element. */
